refactor GenPerspective into its own module and fix fundamental bug
This fixes an issue where when the fly position moved towards an object they
appeared to move further away. The existing perspective logic is correct in a
sense - it was written following
https://csc.lsu.edu/~kooima/articles/genperspective/index.html and correctly
implements what Kooima describes there. But for such a VR cave, the viewing
frustrum gets wider as p_eye approaches the screen, and therefore the perceived
bug was actually correct behavior (for the wrong scenario).

This commit fixes this issue by fixing p_eye and translating the world by the
fly position
